Full theorem deduction
Deducing from traits could cause some duplicate insert failures because
the map would get stale. In general, we could have a race condition here
so shouldn't fail if we try to insert an existing trait.

This commit also adds an admin area with a button to trigger full proof
checking.