Start stratifying module exports
The ultimate goal is to isolate pure logic, db logic, and api definitions, and in general to hide constructors and enforce module boundaries more judiciously.

Next step in this direction is to make sure that the model layer is totally self-encapsulated (e.g. doesn't export `runDB`) and e.g. Persistent could be swapped out if desired.